Extract image endpoint tests to a separate file for parity with audio
